backend/berlin-mcp-server/requirements.txt
mcp>=1.23.0
httpx>=0.27.0
opencensus-ext-azure>=1.1.13
